+++ config/build.js
@@ -4,6 +4,7 @@ import 'babel-polyfill/dist/polyfill.min.js';
 import 'whatwg-fetch';
 import 'riot/riot.min.js';
 import 'moment/min/moment.min.js';
+import 'numeral/min/numeral.min.js';
 
 import H from 'hogan.js/dist/template-3.0.2.min.js';
 window.Hogan = H;